+++ main/loader.c
@@ -304,7 +304,7 @@ static int printdigest(const unsigned char *d)
 	char buf[256]; /* large enough so we don't have to worry */
 
 	for (pos = 0, x = 0; x < 16; x++)
-		pos += sprintf(buf + pos, " %02x", *d++);
+		pos += sprintf(buf + pos, " %02x", (unsigned)*d++);
 
 	ast_debug(1, "Unexpected signature:%s\n", buf);
 
@@ -1205,7 +1205,7 @@ int load_modules(unsigned int preload_only)
 		load_count++;
 
 	if (load_count)
-		ast_log(LOG_NOTICE, "%d modules will be loaded.\n", load_count);
+		ast_log(LOG_NOTICE, "%u modules will be loaded.\n", load_count);
 
 	/* first, load only modules that provide global symbols */
 	if ((res = load_resource_list(&load_order, 1, &modulecount)) < 0) {
